Move DB-dependent policy tests to the integration suite, DEV-20349
