+++ tests/PHPUnit/Unit/Settings/PolicyComparisonTraitTest.php
@@ -4,7 +4,6 @@ namespace Piwik\Tests\Unit\Settings;
 
 use PHPUnit\Framework\TestCase;
 use Piwik\Policy\PolicyEnforcementBypass;
-use Piwik\Policy\PolicyManager;
 use Piwik\Tests\Framework\Mock\Policy\TestPolicy;
 use Piwik\Tests\Framework\Mock\Settings\TraitImpls\PolicyComparisonTraitImpl;
 
@@ -29,46 +28,11 @@ class PolicyComparisonTraitTest extends TestCase
         }
     }
 
-    public function testGetPolicyValuesPolicyInactive()
-    {
-        PolicyManager::setPolicyActiveStatus(TestPolicy::class, false);
-        $values = PolicyComparisonTraitImpl::getPolicyRequiredValues();
-        $this->assertCount(1, $values);
-        $this->assertArrayHasKey(TestPolicy::class, $values);
-        $this->assertNull($values[TestPolicy::class]);
-    }
-
     public function testIsControlledBySpecificPolicy()
     {
         $this->assertTrue(
             PolicyComparisonTraitImpl::isControlledBySpecificPolicy(TestPolicy::class)
         );
-    }
-
-    public function testGetPolicyValuesPresentWhenSettingEnforcementIsEnabledWithoutPolicyFlag()
-    {
-        PolicyManager::setPolicyActiveStatus(TestPolicy::class, false);
-        TestPolicy::setSettingEnforcementSystemValue(PolicyComparisonTraitImpl::class, true);
-
-        try {
-            $values = PolicyComparisonTraitImpl::getPolicyRequiredValues();
-            $this->assertNotNull($values[TestPolicy::class]);
-        } finally {
-            TestPolicy::setSettingEnforcementSystemValue(PolicyComparisonTraitImpl::class, null);
-        }
-    }
-
-    public function testGetPolicyValuesNulledWhenSettingEnforcementIsDisabledDespiteActivePolicyFlag()
-    {
-        PolicyManager::setPolicyActiveStatus(TestPolicy::class, true);
-        TestPolicy::setSettingEnforcementSystemValue(PolicyComparisonTraitImpl::class, false);
-
-        try {
-            $values = PolicyComparisonTraitImpl::getPolicyRequiredValues();
-            $this->assertNull($values[TestPolicy::class]);
-        } finally {
-            TestPolicy::setSettingEnforcementSystemValue(PolicyComparisonTraitImpl::class, null);
-        }
     }
 
     public function testGetPolicyValuesNulledWhileEnforcementIsBypassed()
